[DIA-1352] Wire consent action to new message flow (#554)

+++ cmplibrary/src/main/java/com/sourcepoint/cmplibrary/data/network/model/v7/GetChoiceApiModelExt.kt
@@ -7,8 +7,8 @@ import kotlinx.serialization.json.* //ktlint-disable
 
 internal fun toChoiceBody(
     gdprCs: ConsentStatus?,
-    accountId: Long?,
-    propertyId: Long?,
+    accountId: Int?,
+    propertyId: Int?,
     gdprApplies: Boolean?,
     gdprMessageMetaData: MessageMetaData?,
     fromTest: Boolean = true,
